requirements.txt
asgiref==3.3.1
backcall==0.2.0
decorator==4.4.2
Django==3.1.7
django-bootstrap-form==3.4
django-crispy-forms==1.11.1
django-filter==2.4.0
django-search-listview==0.2.1
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.1
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.0
pytz==2021.1
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
